+++ src/pyrefdev/__main__.py
@@ -1,27 +1,42 @@
-import argparse
 import webbrowser
 import sys
+from typing import Annotated
 
+import cyclopts
+from cyclopts import Parameter
+
+import pyrefdev
 from pyrefdev.mapping import MAPPING
 
 
-def main():
-    parser = argparse.ArgumentParser(
-        prog="pyrefdev",
-        description="pyref.dev is a fast, convenient way to access Python reference docs.",
-    )
-    parser.add_argument("symbol")
-    ns = parser.parse_args()
+app = cyclopts.App("pyrefdev", help=pyrefdev.__doc__)
 
-    symbol = ns.symbol
+
+@app.default
+def main(
+    symbol: str,
+    /,
+    *,
+    should_print: Annotated[
+        bool,
+        Parameter(
+            name=["--print", "-p"],
+            negative="--no-print",
+            help="When true, print the API reference URL instead of opening it in the browser.",
+        ),
+    ] = False,
+):
     if not (url := MAPPING.get(symbol)):
         url = MAPPING.get(symbol.lower())
     if url:
-        webbrowser.open_new_tab(url)
+        if should_print:
+            print(url)
+        else:
+            webbrowser.open_new_tab(url)
     else:
         print(f"{symbol} not found", file=sys.stderr)
         sys.exit(1)
 
 
 if __name__ == "__main__":
-    main()
+    app()